changed variable typedefs, long-id compatibility for better operations with very large (several billion particle) sims. also added option to output positions in double even when rest of file is in signel, also for large sims (where, in zooms with high dynamic range, the difference between positions in code units can become comparable to floating-point precision). finally cleaned up a couple of old unused global variables.

+++ hydro/density.c
@@ -255,10 +255,8 @@ void density(void)
 
   CPU_Step[CPU_DENSMISC] += measure_time();
 
-  int NTaskTimesNumPart;
-
+  long long NTaskTimesNumPart;
   NTaskTimesNumPart = maxThreads * NumPart;
-
   Ngblist = (int *) mymalloc("Ngblist", NTaskTimesNumPart * sizeof(int));
 
   Left = (MyFloat *) mymalloc("Left", NumPart * sizeof(MyFloat));
